# docs/swagger.yaml
definitions:
  types.Column:
    properties:
      exp:
        description: expressions, which default to = when the value is null, have
          =, !=, >, >=, <, <=, like
        type: string
      logic:
        description: logical type, defaults to and when value is null, only &(and),
          ||(or)
        type: string
      name:
        description: column name
        type: string
      value:
        description: column value
    type: object
  types.Conditions:
    properties:
      columns:
        description: columns info
        items:
          $ref: '#/definitions/types.Column'
        type: array
    type: object
  types.CreateUserRequest:
    properties:
      avatarUrl:
        description: 头像
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      likeCount:
        description: 点赞数量
        type: integer
      token:
        type: string
      tokenSecret:
        type: string
      twId:
        description: 推特id
        type: string
      twName:
        description: 推特名
        type: string
      updatedAt:
        type: string
      walletAddress:
        description: 钱包地址
        type: string
    type: object
  types.CreateUserRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          id:
            description: id
            type: integer
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.DeleteUserByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
  types.DeleteUsersByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.DeleteUsersByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
  types.GetUserByConditionRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          user:
            $ref: '#/definitions/types.UserObjDetail'
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.GetUserByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          user:
            $ref: '#/definitions/types.UserObjDetail'
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.ListUsersByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.ListUsersByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          users:
            items:
              $ref: '#/definitions/types.UserObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.ListUsersRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          users:
            items:
              $ref: '#/definitions/types.UserObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.Params:
    properties:
      columns:
        description: query conditions
        items:
          $ref: '#/definitions/types.Column'
        type: array
      page:
        description: page number, starting from page 0
        type: integer
      size:
        description: lines per page
        type: integer
      sort:
        description: sorted fields, multi-column sorting separated by commas
        type: string
    type: object
  types.UpdateUserByIDRequest:
    properties:
      avatarUrl:
        description: 头像
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: uint64 id
        type: integer
      likeCount:
        description: 点赞数量
        type: integer
      token:
        type: string
      tokenSecret:
        type: string
      twId:
        description: 推特id
        type: string
      twName:
        description: 推特名
        type: string
      updatedAt:
        type: string
      walletAddress:
        description: 钱包地址
        type: string
    type: object
  types.UpdateUserByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
  types.UserObjDetail:
    properties:
      avatarUrl:
        description: 头像
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: convert to string id
        type: string
      likeCount:
        description: 点赞数量
        type: integer
      token:
        type: string
      tokenSecret:
        type: string
      twId:
        description: 推特id
        type: string
      twName:
        description: 推特名
        type: string
      updatedAt:
        type: string
      walletAddress:
        description: 钱包地址
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: http server api docs
  title: metax api docs
  version: "2.0"
paths:
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: submit information to create user
      parameters:
      - description: user information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CreateUserRespond'
      security:
      - BearerAuth: []
      summary: create user
      tags:
      - user
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: delete user by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteUserByIDRespond'
      security:
      - BearerAuth: []
      summary: delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: get user detail by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetUserByIDRespond'
      security:
      - BearerAuth: []
      summary: get user detail
      tags:
      - user
    put:
      consumes:
      - application/json
      description: update user information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: user information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdateUserByIDRespond'
      security:
      - BearerAuth: []
      summary: update user
      tags:
      - user
  /api/v1/user/condition:
    post:
      consumes:
      - application/json
      description: get user by condition
      parameters:
      - description: query condition
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Conditions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetUserByConditionRespond'
      security:
      - BearerAuth: []
      summary: get user by condition
      tags:
      - user
  /api/v1/user/delete/ids:
    post:
      consumes:
      - application/json
      description: delete users by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.DeleteUsersByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteUsersByIDsRespond'
      security:
      - BearerAuth: []
      summary: delete users
      tags:
      - user
  /api/v1/user/list:
    get:
      consumes:
      - application/json
      description: list of users by last id and limit
      parameters:
      - default: 0
        description: last id, default is MaxInt32
        in: query
        name: lastID
        required: true
        type: integer
      - default: 10
        description: size in each page
        in: query
        name: limit
        type: integer
      - default: -id
        description: 'sort by column name of table, and the '
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListUsersRespond'
      security:
      - BearerAuth: []
      summary: list of users by last id and limit
      tags:
      - user
    post:
      consumes:
      - application/json
      description: list of users by paging and conditions
      parameters:
      - description: query parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Params'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListUsersRespond'
      security:
      - BearerAuth: []
      summary: list of users by query parameters
      tags:
      - user
  /api/v1/user/list/ids:
    post:
      consumes:
      - application/json
      description: list of users by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.ListUsersByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListUsersByIDsRespond'
      security:
      - BearerAuth: []
      summary: list of users by batch id
      tags:
      - user
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type Bearer your-jwt-token to Value
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"